requirements.txt
gradio
requests
beautifulsoup4
lxml
pandas
tldextract
textstat
xlsxwriter
openpyxl
playwright
pillow
google-generativeai
python-dotenv
